% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyDictFun.R
\name{applyDict}
\alias{applyDict}
\title{apply dict function}
\usage{
applyDict(
  column.name = "GROUP",
  dict.name = "main",
  column.event.name = "EVENT",
  pat.process,
  param.EVENTName = "EVENT",
  df.groups
)
}
\description{
apply dict function
}
