% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_KM_fun.R
\name{combine_km_fun}
\alias{combine_km_fun}
\title{creo struttura per combinare tutti i plot delle km}
\usage{
combine_km_fun(list.path, path.plot)
}
\description{
creo struttura per combinare tutti i plot delle km
}
