% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_time_fun.R
\name{cov_time_fun}
\alias{cov_time_fun}
\title{function for cov analysis}
\usage{
cov_time_fun(
  ObjDL,
  ObjCFM,
  covariate,
  arr.from,
  lst.to,
  covariate.type = "attribute",
  is.numerical = TRUE,
  abs.threshold = NA,
  UM = "days",
  plot.points = TRUE,
  plot.RegressionLine = FALSE,
  points.symbols = 20,
  size.symbols = 1.5,
  line.width = 2,
  y.int.legend = 0.6,
  legend.text.size = 0.6,
  legend.position = "topleft"
)
}
\description{
function for cov analysis
}
