% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_pat_fun.R
\name{event_pat}
\alias{event_pat}
\title{Event distribution over patients}
\usage{
event_pat(data, chosen, flag)
}
\description{
this function creates a bar plot that shows the distribution of events by patient,
showing how many patients shared a specific type of event.
}
