% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_fun.R
\name{heatmap}
\alias{heatmap}
\title{Co-occurrence matrix}
\usage{
heatmap(data, chosen, objDL.new)
}
\description{
this function creates a matrix that shows the co-coccurrence of the selected events,
calculated as the number of times (among the total) each event occurs in the presence of another,
within the clinical path of the same patient
}
