% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM_CFM.R
\name{KM_CFM}
\alias{KM_CFM}
\title{KM function x CFM}
\usage{
KM_CFM(
  ObjCFM,
  id.start,
  id.end,
  cens.leaf = TRUE,
  id.cens = c(),
  ObjDL,
  UM = "mins",
  min_time = 0,
  max_time = Inf,
  id_strat = c()
)
}
\description{
KM function x CFM
}
