% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM_FOMM.R
\name{KM_FOMM}
\alias{KM_FOMM}
\title{KM function}
\usage{
KM_FOMM(
  fromState,
  toState,
  ObjDL,
  passingThrough = c(),
  passingNotThrough = c(),
  stoppingAt = c(),
  stoppingNotAt = c(),
  PDVAt = c(),
  withPatientID = c(),
  UM = "mins"
)
}
\description{
KM function
}
