% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_km_graph_FOMM.R
\name{render.km.graph.FOMM}
\alias{render.km.graph.FOMM}
\title{creo struttura per combinare tutti i plot delle km}
\usage{
render.km.graph.FOMM(list.path, path.plot)
}
\description{
creo struttura per combinare tutti i plot delle km
}
