% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_km_function.R
\name{render.km.graph}
\alias{render.km.graph}
\title{creo struttura per combinare tutti i plot delle km}
\usage{
render.km.graph(list.path, path.plot)
}
\description{
creo struttura per combinare tutti i plot delle km
}
