% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_new_rule_mod.R
\name{spec_setting_ui}
\alias{spec_setting_ui}
\title{Merge module: Merge by attribute specific rules settings}
\usage{
spec_setting_ui(id, var.name, var_long_type)
}
\description{
Merge module: Merge by attribute specific rules settings
}
