% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_dist_fun.R
\name{time_dist}
\alias{time_dist}
\title{Event distribution over time}
\usage{
time_dist(data, chosen, date_range)
}
\description{
this function creates a bar plot The bar plot that shows the distribution
of the  events in different time.the time axis has been discretized using quarters,
in order to highlight how the distribution of each event changes over time. it is also possible to
set a time window in order to analize event distribution in a smaller temporal range.
}
