% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_query.R
\name{trace.id}
\alias{trace.id}
\title{trace id}
\usage{
trace.id(
  Obj.QOD,
  event.start,
  event.end,
  time.b,
  time.a,
  inf.flag,
  um.time,
  event.between,
  comp.mat = TRUE
)
}
\description{
trace id
}
