% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_long.R
\name{ui_mod_long}
\alias{ui_mod_long}
\title{Merge module: page for ancillary data recap and both merging techniques}
\usage{
ui_mod_long(id, data, num)
}
\description{
Merge module: page for ancillary data recap and both merging techniques
}
