% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Upset_fun.R
\name{upset_fun}
\alias{upset_fun}
\title{Upset plot}
\usage{
upset_fun(data, chosen, objDL.new)
}
\description{
this function creates an upset plot in order to analyze the co-occurrence of the events
by showing the frequencies of the combination sets.
}
